// apps/web/lib/api/client/models/Patient.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Patient = {
    patient_id: string;
    patient_fname: string;
    patient_mname?: (string | null);
    patient_lname: string;
    patient_dob: string;
    patient_contact_phone: string;
    patient_contact_email: string;
    patient_gender: string;
    patient_insurance_member_id?: (string | null);
};